Improved camera stream efficiency (#37)
For every camera stream widget, instead of having a separate http
request and stream each time it is rendered, it uses data from the same
stream. This also allows for auto disposal/closing whenever the widget
becomes invisible, and the actual camera stream widget doesn't need to
be cached, only the state has to be.