Prevent simultaneous connections (#202)
Previously, if the IP address was changed while network tables was
disconnected, there would be a new connection attempt "loop" that tried
to connect at the same time as the existing connection attempt
"loop(s)". When the connection was eventually established, multiple
clients would be created on Network Tables.

Now, when trying to connect, any additional calls to `_connect()` while
trying to connect will be blocked off, preventing this issue